pyproject.toml
[tool.poetry]
name = "automastergraphqlapi"
version = "0.1.0"
description = ""
authors = ["who-man-tech <wwho.mann.3@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
graphene-django = "^2.15.0"
psycopg2-binary = "^2.9.3"
Pillow = "^9.0.1"
django-cors-headers = "^3.11.0"
Django = "^3.2.12"
argon2-cffi = "^21.3.0"
django-graphql-jwt = "^0.3.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
